+++ src/plugins.ts
@@ -1,0 +1,10 @@
+import layout from "./plugins/layout";
+import performance from "./plugins/performance";
+import pointer from "./plugins/pointer";
+import viewport from "./plugins/viewport";
+
+const classes = { layout, viewport, pointer, performance };
+
+export function plugins(name) {
+  return classes[name];
+}
